[Refactor] 프로젝트 전체 코드 리팩토링
- FoodController 내 URI 중복 해결
- Foot 엔티티 내 toResponse 메서드 일부 수정
- PurchaseController, PurchaseService 내에 투표(Vote) 관련 메서드를 모두 VoteController와 VoteService로 이동
