+++ src/main/kotlin/team/b2/bingojango/domain/purchase/controller/PurchaseController.kt
@@ -1,13 +1,10 @@
 package team.b2.bingojango.domain.purchase.controller
 
 import io.swagger.v3.oas.annotations.Operation
-import jakarta.validation.Valid
-import org.apache.coyote.Response
 import org.springframework.http.ResponseEntity
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.*
 import team.b2.bingojango.domain.purchase.service.PurchaseService
-import team.b2.bingojango.domain.vote.dto.request.VoteRequest
 import team.b2.bingojango.global.security.UserPrincipal
 
 @RestController
@@ -50,31 +47,4 @@ class PurchaseController(
         @PathVariable refrigeratorId: Long
     ) =
         purchaseService.showPurchase(refrigeratorId)
-
-    @Operation(summary = "현재 공동구매 목록에 대한 투표 현황 조회")
-    @GetMapping("/vote/{voteId}")
-    fun showVote(
-        @PathVariable refrigeratorId: Long,
-        @PathVariable voteId: Long
-    ) =
-        ResponseEntity.ok().body(purchaseService.showVote(refrigeratorId, voteId))
-
-    @Operation(summary = "현재 공동구매 목록에 대한 투표 시작")
-    @PostMapping("/vote")
-    fun startVote(
-        @AuthenticationPrincipal userPrincipal: UserPrincipal,
-        @PathVariable refrigeratorId: Long,
-        @Valid @RequestBody voteRequest: VoteRequest
-    ) =
-        ResponseEntity.ok().body(purchaseService.startVote(userPrincipal, refrigeratorId, voteRequest))
-
-    @Operation(summary = "현재 공동구매 목록에 대한 투표")
-    @PutMapping("/vote/{voteId}")
-    fun vote(
-        @AuthenticationPrincipal userPrincipal: UserPrincipal,
-        @PathVariable refrigeratorId: Long,
-        @PathVariable voteId: Long,
-        @RequestParam isAccepted: Boolean
-    ) =
-        ResponseEntity.ok().body(purchaseService.vote(userPrincipal, refrigeratorId, voteId, isAccepted))
 }
